+++ yastn/tensor/_tests.py
@@ -18,7 +18,7 @@ from operator import mul
 
 import numpy as np
 
-from ._auxliary import _flatten, _unpack_axes, _struct
+from ._auxiliary import _flatten, _unpack_axes, _struct
 
 __all__ = ['are_independent', 'is_consistent', 'YastnError']
 
@@ -31,17 +31,11 @@ def _test_can_be_combined(a, b):
     """Check if config's of two tensors allow for performing operations mixing them. """
     if type(a) is not type(b):
         raise YastnError('Operation requires two yastn.Tensor-s')
-    if a.device != b.device:
-        raise YastnError('Devices of the two tensors do not match.')
-    _test_configs_match(a.config, b.config)
-
-
-def _test_configs_match(a_config, b_config):
-    if a_config.sym.SYM_ID != b_config.sym.SYM_ID:
+    if a.config.sym.SYM_ID != b.config.sym.SYM_ID:
         raise YastnError('Two tensors have different symmetry rules.')
-    if a_config.fermionic != b_config.fermionic:
+    if a.config.fermionic != b.config.fermionic:
         raise YastnError('Two tensors have different assigment of fermionic statistics.')
-    if a_config.backend.BACKEND_ID != b_config.backend.BACKEND_ID:
+    if a.config.backend.BACKEND_ID != b.config.backend.BACKEND_ID:
         raise YastnError('Two tensors have different backends.')
 
 
@@ -53,22 +47,23 @@ def _test_tD_consistency(struct):
         Di = Dset[:, i].tolist()
         tDi = list(zip(ti, Di))
         if len(set(ti)) != len(set(tDi)):
-            raise YastnError('Inconsist assigment of bond dimension to some charge.')
+            raise YastnError('Inconsistent assignment of bond dimension to some charge.')
 
 
-def _test_axes_match(a, b, sgn=1, axes=None):
+def _unpack_trans_test_axes_pair(a, b, sgn=1, axes=None):
     """
     Test if legs of a in axes[0] and legs ob b in axes[1] have matching signature and fusion structures.
     sgn in (-1, 1) is the sign of required match between signatures.
 
-    Return meta-unfused axes and information if hard-fusion mask will be required.
+    Return meta-unfused axes (after application of trans) and information if hard-fusion mask will be required.
     """
 
     if axes is None:
         if a.ndim != b.ndim:
             raise YastnError('Tensors have different number of legs.')
         axes = (tuple(range(a.ndim)), tuple(range(b.ndim)))
-        uaxes = (tuple(range(a.ndim_n)), tuple(range(b.ndim_n)))
+        ua = tuple(range(a.ndim_n))
+        ub = tuple(range(b.ndim_n))
     else:
         if len(axes[0]) != len(axes[1]):
             raise YastnError('axes[0] and axes[1] indicate different number of legs.')
@@ -79,23 +74,26 @@ def _test_axes_match(a, b, sgn=1, axes=None):
             raise YastnError('Axis outside of tensor ndim.')
         ua, = _unpack_axes(a.mfs, axes[0])
         ub, = _unpack_axes(b.mfs, axes[1])
-        uaxes = (ua, ub)
 
-    if not all(a.struct.s[i1] == sgn * b.struct.s[i2] for i1, i2 in zip(*uaxes)):
+    # transpose axes
+    haxes = (tuple(a.trans[ax] for ax in ua),
+             tuple(b.trans[ax] for ax in ub))
+
+    if not all(a.struct.s[i1] == sgn * b.struct.s[i2] for i1, i2 in zip(*haxes)):
         raise YastnError('Signatures do not match.')
 
     if any(a.mfs[i1] != b.mfs[i2] for i1, i2 in zip(*axes)):
         raise YastnError('Indicated axes of two tensors have different number of meta-fused legs or sub-fusions order.')
 
     mask_needed = False  # for hard-fused legs
-    for i1, i2 in zip(*uaxes):
+    for i1, i2 in zip(*haxes):
         if a.hfs[i1].tree != b.hfs[i2].tree or a.hfs[i1].op != b.hfs[i2].op:
             raise YastnError('Indicated axes of two tensors have different number of hard-fused legs or sub-fusions order.')
         if any(s1 != sgn * s2 for s1, s2 in zip(a.hfs[i1].s, b.hfs[i2].s)):
             raise YastnError('Signatures of hard-fused legs do not match.')
         if a.hfs[i1].t != b.hfs[i2].t or a.hfs[i1].D != b.hfs[i2].D:
             mask_needed = True
-    return mask_needed, uaxes
+    return mask_needed, haxes
 
 
 def _test_axes_all(a, axes, native=False):
@@ -118,7 +116,7 @@ def are_independent(a, b, independent=True):
 
 def is_consistent(a):
     """
-    Test is yastn tensor is does not contain inconsistent structures
+    Test if yastn tensor does not contain inconsistent structures
 
     Check that:
     1) tset and Dset correspond to A
@@ -172,7 +170,6 @@ def _test_struct_types(struct):
     assert all(isinstance(y, int) for x in struct.D for y in x)
     assert isinstance(struct.D, tuple)
     assert isinstance(struct.size, int)
-
 
 
 def _get_tD_legs(struct):