+++ CHANGELOG.md
@@ -48,7 +48,7 @@
 - fix passing dtype to `backend.to_tensor`. The latter always uses specified dtype (without try ... except ...)
 
 04-06-2021
-- new function `add_leg` which addes an auxliary leg with dimension 1, carrying a charge of the tensor (or part of it)
+- new function `add_leg` which addes an auxiliary leg with dimension 1, carrying a charge of the tensor (or part of it)
 
 07-06-2021
 - option `reverse` added to `to_dense()`, `to_numpy()` and `to_nonsymmetric()` functions allowing for reverse